[Benja, Dani C,  Dani S, Bere] fixing  symptoms registry tests and refactoring
