+++ package.json
@@ -41,6 +41,7 @@
     "@chakra-ui/core": "^0.8.0",
     "@emotion/core": "^10.0.35",
     "@emotion/styled": "^10.0.27",
+    "axios": "^0.21.0",
     "emotion-theming": "^10.0.27",
     "next": "latest",
     "pg": "^8.4.1",
