+++ pages/api/registry-save.ts
@@ -3,14 +3,16 @@ import { NextApiRequest, NextApiResponse } from 'next'
 import { RegistryService } from 'src/services/RegistryService'
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  const symptoms = JSON.parse(req.body)
+  const symptoms = req.body
   try {
     const registryService = new RegistryService()
-    registryService.saveRegistry(symptoms)
+    await registryService.saveRegistry(symptoms)
+    res.status(200)
   } catch (err) {
     console.error(err)
-    res.statusCode = 500
+    res.status(500)
   }
+  res.send(null)
 }
 
 export default handler
