+++ test/__endpoints/registry-request.http
@@ -1,4 +1,4 @@
-POST http://localhost:3000/api/symptom HTTP/1.1
+POST http://localhost:3000/api/registry-save HTTP/1.1
 content-type: application/json
 
 {
